Remove a dead function, and make every dependency explicit
acceptAliasDigest had no caller and could not have the one its comment imagined:
wallets sign typed data, not bare digests. The encoding it was meant to pin is
already proven end to end — e2e-live builds a signature with the SDK's own
acceptAlias and submits it, which fails if the two disagree. That is a stronger
check than digest equality, and it does not cost permanent bytecode in a contract
that cannot be changed.

Swept the rest of the public surface for the same thing. Everything else with no
callers is inherited ERC-721 and EIP-712 (getApproved, tokenURI, supportsInterface,
eip712Domain) which cannot be removed, or getters that exist so a reader can
verify a wiring — transactVerifier, registryRootSeenAt, the level constants. Those
stay.

Then a stricter solhint: 90 warnings down to 46, none of them errors. Thirty
global imports became named ones, and the first compile after that failed — which
was the point. HaliasRegistry had been using PoseidonT4 and FIELD_PRIME that
arrived transitively through SMTRegistry's own global import, so nothing in the
file said where they came from. They are direct now.

The other fourteen were `uint[2]` mirroring the shape snarkjs generates. `uint` is
an alias for `uint256`, so making them explicit is type-identical and the verifier
call is unaffected.

Left as warnings deliberately: gas-increment-by-one and gas-strict-inequalities
suggest rewriting comparisons for savings the optimiser mostly already takes,
immutable-vars-naming would rename controller/pool/registry for style, and
gas-struct-packing cannot apply to TransactParams — its layout is the ABI a proof
commits to.

174 hardhat, 65 sdk, 108 e2e-live.

+++ packages/protocol/contracts/interfaces/ITransactVerifier.sol
@@ -3,9 +3,9 @@ pragma solidity 0.8.28;
 
 interface ITransactVerifier {
     function verifyProof(
-        uint[2] calldata _pA,
-        uint[2][2] calldata _pB,
-        uint[2] calldata _pC,
-        uint[14] calldata _pubSignals
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[14] calldata _pubSignals
     ) external view returns (bool);
 }
